[feat]: finish LineBlock
added press menu again
removed reduntant code
